requirements.txt
numpy==1.21.4
python-dotenv==0.19.2
requests==2.26.0
bs4
